+++ sdk/ai/Azure.AI.Projects.Agents/CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ### Sample Updates
 - Added sample demonstrating disabling and enabling Hosted Agent.
-- Added samples for Agent optinmization jobs.
+- Added samples for Agent optimization jobs.
 
 ## 2.1.0-beta.4 (2026-06-30)