+++ sdk/ai/Azure.AI.Projects/CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Breaking Changes
 
 ### Bugs Fixed
-- Fix the Feature-Foundry header, resulting in 400 error during Agent optimization jobs submission.
+- Fix the Foundry-Features header, resulting in 400 error during Agent optimization jobs submission.
 
 ### Other Changes